Progress: Add dynamic, dynamics, and map unit tests for fl_fss_payload_header_map() and fix several problems.
Switch character width loops with memcpy().

Add cases where trailing string should be removed.

Make sure private_fl_fss_basic_write() is only called when the cache is not empty.
Make sue the internal.range is properly set when calling private_fl_fss_basic_write().

Ensure that when any quote is used that they get quoted where appropriate.

Ensure that when key has used of 0, the start range is greater than the stop range to designate a NULL range.

Add dynamic, dynamics, and map unit tests.

Make the tests test the string first rather than the used so that when there is an error it is easier to debug.

Remove unused/stale commented out code.

Switch fss_write flag from 16-bit into 32-bit.
This is necessary because more bits than 16 are in use.

+++ level_3/fss_write/c/main/common/type.h
@@ -81,7 +81,7 @@ extern "C" {
  */
 #ifndef _di_fss_write_setting_t_
   typedef struct {
-    uint16_t flag;
+    uint32_t flag;
 
     f_status_t status_signal;
     f_state_t state;
